% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{aplot.brewerSets}
\alias{aplot.brewerSets}
\title{Helper function to stick RColorBrewer palettes Set1, Set2, and Set3 together
and remove the bright yellow colors. If you provide a value for \code{n} then
it will return that many colors starting from the first. If the value of \code{n}
is longer than the total length of the colors then the colors are interpolated
with \link{colorRampPalette}.}
\usage{
aplot.brewerSets(n = NULL)
}
\arguments{
\item{n}{Number of colors to return. If \code{n == NULL} then all of the colors are returned.}
}
\value{
A vector of colors
}
\description{
Helper function to stick RColorBrewer palettes Set1, Set2, and Set3 together
and remove the bright yellow colors. If you provide a value for \code{n} then
it will return that many colors starting from the first. If the value of \code{n}
is longer than the total length of the colors then the colors are interpolated
with \link{colorRampPalette}.
}
