% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{aplot.statErr}
\alias{aplot.statErr}
\title{Calculate mean/median and error of values in vector \code{x}}
\usage{
aplot.statErr(x, fun = c("mean", "median"), err.fun = c("sd", "sem", "var",
  "ci"), ci = 0.95, num.bootstraps = 1, boot.err.mode = 1,
  groups = NULL, grand.stat = FALSE)
}
\arguments{
\item{x}{Vector of numbers}

\item{fun}{Statistic. Either \code{mean} or \code{median}. Default is \code{mean}}

\item{err.fun}{Error/uncertainty to calculate. May be \code{sd}, \code{sem}, \code{var}, or \code{ci}. \code{ci} returns the quantiles of the input values unless you are bootstrapping in which case the confidence interval of the bootstrap distribution is returned.}

\item{ci}{Confidence interval to report when using \code{err.fun="ci"}}

\item{num.bootstraps}{Number of boostrap iterations. If \code{num.bootstraps == 1} then no bootstrapping is performed.}

\item{boot.err.mode}{Set to either 1, default or 2. Default behavior is to apply the
error function, \code{err.fun}, to the bootstrap distribution. When set to 2
the error is bootstrapped and the mean of it's bootstrap distribution is returned.}

\item{groups}{Grouping vector for values in \code{x}. If provided then this function operates on each group of values and returns a matrix of results.}

\item{grand.stat}{If using \code{groups} this calculates the overall statistic and uncertainty by first calculating the group statistic and uncertainties then propagating that into the grand statistic.}
}
\value{
A vector of values or a matrix if \code{groups} was specified.
}
\description{
Calculate mean/median and error of values in vector \code{x}
}
