% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{xlab2}
\alias{xlab2}
\alias{ylab2}
\title{display x or y axis label as an ordinary text, so that the label will not be aligned with axis label of another plot}
\usage{
xlab2(label, fontsize = 11, ...)

ylab2(label, fontsize = 11, ...)
}
\arguments{
\item{label}{axis label}

\item{fontsize}{fontsize of the label}

\item{...}{additional parameter passed to `gpar`}
}
\value{
gg object with new label
}
\description{
display x or y axis label as an ordinary text, so that the label will not be aligned with axis label of another plot
}
\author{
Guangchuang Yu
}
