% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplot.R
\name{as.patchwork}
\alias{as.patchwork}
\title{as.patchwork}
\usage{
as.patchwork(x, modify_xlim = TRUE, modify_ylim = TRUE)
}
\arguments{
\item{x}{object}

\item{modify_xlim}{use xlim2() to modify xlim or not}

\item{modify_ylim}{use ylim2() to modify ylim or not}
}
\description{
as.patchwork
}
