% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap.R
\name{funky_bar}
\alias{funky_bar}
\title{funky_bar}
\usage{
funky_bar(data, cols)
}
\arguments{
\item{data}{data frame}

\item{cols}{selected columns}
}
\value{
ggplot object
}
\description{
create bar plot for funkyheatmap
}
\author{
Guangchuang Yu
}
