% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap.R
\name{funky_heatmap}
\alias{funky_heatmap}
\title{funky_heatmap}
\usage{
funky_heatmap(..., data = NULL, widths = NULL, options = NULL)
}
\arguments{
\item{...}{funky plots (e.g., outputs of `funky_point`, `funky_bar`, etc.)}

\item{data}{If data is provided, create a funkyheatmap from it. Otherwise, create composite plot from `...`}

\item{widths}{relative widths of the plots}

\item{options}{any ggplot component that can be added to the plots}
}
\value{
gglist object
}
\description{
create a funkyheatmap
}
\author{
Guangchuang Yu
}
