% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap.R
\name{funky_text}
\alias{funky_text}
\title{funky_text}
\usage{
funky_text(data, cols = 1, hjust = 0)
}
\arguments{
\item{data}{data frame}

\item{cols}{selected column}

\item{hjust}{text alignment adjustment}
}
\value{
ggplot object
}
\description{
create text plot (i.e., rownames) for funkyheatmap
}
\author{
Guangchuang Yu
}
