% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{get_all_subsets}
\alias{get_all_subsets}
\title{Get the items/names/ids of subsets from a named list}
\usage{
get_all_subsets(list, name_separator = "/")
}
\arguments{
\item{list}{a named list}

\item{name_separator}{default is /}
}
\value{
a tibble
}
\description{
Get the items/names/ids of subsets from a named list
}
\examples{
list = list(A = sample(LETTERS, 20),
            B = sample(LETTERS, 22),
            C = sample(LETTERS, 24),
            D = sample(LETTERS, 30, replace = TRUE))
get_all_subsets(list)
}
