% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-list.R
\name{gglist}
\alias{gglist}
\title{construct a `gglist` object}
\usage{
gglist(gglist, ...)
}
\arguments{
\item{gglist}{a list of plots}

\item{...}{parameters for plotting the `gglist`}
}
\value{
gglist object
}
\description{
constructure a `gglist` object that contains a list of plots (`gglist`) and 
parameters (via `...`), the object can be displayed via the `plot_list()` function.
}
\author{
Guangchuang Yu
}
