% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-list.R
\name{gglistGrob}
\alias{gglistGrob}
\title{This function converts 'gglist' object to grob (i.e. gtable object)}
\usage{
gglistGrob(x)
}
\arguments{
\item{x}{A 'gglist' object}
}
\value{
A 'gtable' object
}
\description{
title gglistGrob
}
