% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{yrange}
\alias{yrange}
\alias{xrange}
\title{plot range of a ggplot object}
\usage{
yrange(gg)

xrange(gg)
}
\arguments{
\item{gg}{a ggplot object}
}
\value{
range of selected axis
}
\description{
extract x or y ranges of a ggplot
}
\author{
Guangchuang Yu
}
