% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.r
\name{oncoplot}
\alias{oncoplot}
\title{ploting oncoplot with aplot}
\usage{
oncoplot(maf, genes = 20)
}
\arguments{
\item{maf}{MAF object.}

\item{genes}{the gene names or the number, default is 20.}
}
\value{
\code{oncoplot} object, which is also a \code{aplot} object
}
\description{
ploting oncoplot with aplot
}
\examples{
\dontrun{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml.clin <- system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
laml <- maftools::read.maf(maf = laml.maf, clinicalData = laml.clin)
oncoplot(maf = laml, genes = 20)
}
}
