% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-list.R
\name{plot_list}
\alias{plot_list}
\title{plot a list of ggplot objects}
\usage{
plot_list(
  ...,
  gglist = NULL,
  ncol = NULL,
  nrow = NULL,
  widths = NULL,
  heights = NULL,
  tag_levels = NULL,
  tag_size = 14
)
}
\arguments{
\item{...}{list of plots to be arranged}

\item{gglist}{(optional) list of plots}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{widths}{relative widths}

\item{heights}{relative heights}

\item{tag_levels}{format to label plots}

\item{tag_size}{size of tags}
}
\value{
composite plot
}
\description{
plot a list of ggplot objects using patchwork, similar to `cowplot::plot_grid(plotlist)`
}
\author{
Guangchuang Yu
}
