% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggsave}
\alias{\%>\%}
\alias{ggrange}
\alias{xrange}
\alias{yrange}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggfun}{\code{\link[ggfun:ggrange]{ggrange()}}, \code{\link[ggfun:xrange]{xrange()}}, \code{\link[ggfun:yrange]{yrange()}}}

  \item{ggplot2}{\code{\link[ggplot2:ggsave]{ggsave()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}}

