% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upsetplot}
\alias{upsetplot}
\title{upsetplot}
\usage{
upsetplot(
  list,
  nintersects = 40,
  order.intersect.by = c("size", "name"),
  order.set.by = c("size", "name")
)
}
\arguments{
\item{list}{a list of sets}

\item{nintersects}{number of intersects}

\item{order.intersect.by}{one of 'size' or 'name'}

\item{order.set.by}{one of 'size' or 'name'}
}
\value{
an upset plot
}
\description{
Plot upsetplot
}
\details{
This function generate upsetplot by creating a composite plot which contains subplots genreated by ggplot2
}
