% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplot-package.R
\docType{package}
\name{aplot-package}
\alias{aplot}
\alias{aplot-package}
\title{aplot: Decorate a 'ggplot' with Associated Information}
\description{
'aplot' supports data-driven composition of a main plot with associated subplots that need precise axis alignment. Unlike general layout-focused tools such as 'cowplot' and 'patchwork', it enables related subplots to be integrated on the top, bottom, left, or right sides of a main plot with matched axes, so that the combined panels can be interpreted as a coherent whole. This design was inspired by the 'Method 2' described in 'ggtree' (G Yu (2018) \doi{10.1093/molbev/msy194}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/aplot}
  \item \url{https://yulab-smu.top/aplot/}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})
}

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
  \item Thomas Hackl \email{thackl@mit.edu} [contributor]
}

}
\keyword{internal}
