% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplot-package.R
\docType{package}
\name{aplot-package}
\alias{aplot}
\alias{aplot-package}
\title{aplot: Decorate a 'ggplot' with Associated Information}
\description{
For many times, we are not just aligning plots as what 'cowplot' and 'patchwork' did. Users would like to align associated information that requires axes to be exactly matched in subplots, e.g. hierarchical clustering with a heatmap. This package provides utilities to aligns associated subplots to a main plot at different sides (left, right, top and bottom) with axes exactly matched.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/aplot}
  \item \url{https://yulab-smu.top/aplot/}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
  \item Thomas Hackl \email{thackl@mit.edu} [contributor]
}

}
\keyword{internal}
