% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplot.R
\name{as.patchwork}
\alias{as.patchwork}
\title{as.patchwork}
\usage{
as.patchwork(x, align = getOption("aplot_align", default = "xy"))
}
\arguments{
\item{x}{object}

\item{align}{"x","y","xy","none", align the axis of x/y or not.}
}
\description{
as.patchwork
}
