% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap.R
\name{funky_point}
\alias{funky_point}
\title{funky_point}
\usage{
funky_point(data, cols, cols2 = NULL, ...)
}
\arguments{
\item{data}{data frame}

\item{cols}{selected columns}

\item{cols2}{selected columns to keep names}

\item{...}{additional parameters, passing to \code{\link[ggstar]{geom_star}}}
}
\value{
ggplot object
}
\description{
create dot plot for funkyheatmap
}
\author{
Guangchuang Yu
}
