% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-list.R
\name{plot_list}
\alias{plot_list}
\title{plot a list of ggplot objects}
\usage{
plot_list(
  ...,
  gglist = NULL,
  ncol = NULL,
  nrow = NULL,
  byrow = NULL,
  widths = NULL,
  heights = NULL,
  guides = NULL,
  labels = NULL,
  tag_levels = NULL,
  tag_size = 14,
  design = NULL
)
}
\arguments{
\item{...}{list of plots to be arranged}

\item{gglist}{(optional) list of plots}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{byrow}{If "FALSE" the plots will be filled in in column-major order}

\item{widths}{relative widths}

\item{heights}{relative heights}

\item{guides}{A string specifying how guides should be treated in the layout.}

\item{labels}{manual specified labels to label plots}

\item{tag_levels}{format to label plots, will be disable if 'labels' is not NULL}

\item{tag_size}{size of tags}

\item{design}{specification of the location of areas in the layout}
}
\value{
composite plot
}
\description{
plot a list of ggplot objects using patchwork, similar to `cowplot::plot_grid(plotlist)`
}
\author{
Guangchuang Yu
}
