% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-list.R
\name{plot_list}
\alias{plot_list}
\title{plot a list of ggplot objects}
\usage{
plot_list(gglist, ncol = NULL, nrow = NULL, widths = NULL, heights = NULL, ...)
}
\arguments{
\item{gglist}{list of ggplot objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{widths}{relative widths}

\item{heights}{relative heights}

\item{...}{additional parameters that passed to plot_layout}
}
\value{
composite plot
}
\description{
plot a list of ggplot objects using patchwork, similar to `cowplot::plot_grid(plotlist)`
}
\author{
Guangchuang Yu
}
